Добавление шаблонов к заказам и незначительные корректировки.
- В настройках удаление лишних строчек и настройка статики
- В админке задано отоброжение для пустых полей для всего приложения
- В модели заказа изменение поля на вещественное число
- В представлениях добавлены отображения детальной информации заказа и обновление главной страницы
- Подключение статических файлов и создание их
- Корректировки навигации в шаблонах

+++ kraft/production/models.py
@@ -37,7 +37,7 @@ class Order(models.Model):
     """
     name = models.CharField('Имя', max_length=MAX_LENGTH_NAME)
     count = models.IntegerField('Количество')
-    weight_one_piece = models.IntegerField(
+    weight_one_piece = models.FloatField(
         'вес одной шт. в граммах', null=True
     )
     package_size = models.CharField(
